Set FCMASK in hints file.
By default, the file mask for files generated like the dumplog, is 0660.
For a public server trying to share these publically on the web, this
file mask won't allow the public to read the file. We've been running a
script server-side to correct these file masks for a long time now,
seems that shouldn't have been necessary.