Disable experimental TTY_PERM_INVENT.
Disabling this in config.h, currently it looks like crap, plus the
warning message it will display at the start of each game that doesn't
meet the size criteria is going to confuse a lot of players.

Easy enough to re-enable via hints file once it's out of experimental
status.