htmldump - fix crash that occurred from death by wand zap in tty windowport
This is a partial fix - do_screen_desc/lookat were crashing when trying to
get the description of the zap beam for the html dump.  This changes
glyph_to_cmap to include zap glyphs (which prevents the crash) but there is
some issue in do_screen_desc that causes it to be described as "interior of
a monster", which is a problem for another day.