move former_rank[] out of monst, into mextra
I'd like to have save and bones file compatibility between
hardfought 3.7 and the upstream NetHack-3.7 work-in-progress.
There are two hardfought additions that render things incompatible
as they currently stand:

1) the elbereth field addition to u_conduct. That one is
easily handled on the NetHack-3.7-side by us and requires
no change on the hardfought side. A placeholder field can
be added at the same place (For now, it will probably be
called hf_reserved1 or something like that on the upsteam
NetHack-3.7, since there won't be any code support for it
there, at the time it is added anyway. You would just
leave it the way it is on the hardfought-side. Adjust to
rename the new field on hardfought (not add an additional
field) the next time you sync with upsteam NetHack-3.7
after these changes..

2) This field that hardfought adds to include/monst.h:
char former_rank[25];  /* for bones' ghost rank in their former life */

I'd like to propose handling that former_rank in a different
way. Instead of housing it in the base monst struct as hardfought
has it now (which makes it part of every monst in the game,
whether it is needed or not), I'd like to relocate it into a
new mextra struct.

This pull-request does just that. It adds an additional field
called 'former' to the mextra struct, and adds supporting code
to deal with it for the specific situation where it is needed.
If the pull-request is accepted into hardfought 3.7, I can add
similar functionality to upsteam NetHack-3.7 and once both
places are changed, the savefiles should be compatible.

That will make it simpler when a savefile needs to be loaded
while trying to debug and issue or bug report going forward.

I have the upsteam NetHack-3.7 change standing by once the
fate of this pull request is known.

The changes, once incorporated, will not be compatible with
existing save and bones files that preceded the changes.