Rate-limit character rerolling.
I've added this throttling because it was discovered early on that if a
player were to ride the reroll key, it would cause a massive spike in
CPU usage of the binary. A member of the devteam stated that a fix on
their end needs to encompass all platofrms the game is supported on. I
don't want to wait that long - this commit adds a one second throttle to
rerolling on unix/linux/bsd/macos/windows platforms.